Add files via upload
Among all the classifiers, it was found that XGBoost (eXtreme Gradient Boosting) provided us with an accuracy of 93.333%.Based on this we recommend that the XGBoost technique should be used to develop a model for Parkinson’s disease detection problems. 
Even though the model works efficiently, this is limited to the richness of the dataset that was used to train this model. 
The selected dataset that was used in this study had 195 instances, hence a dataset with more instances would help to generalize better. This proposed model is a reliable model to detect Parkinson’s Disease due to its high accuracy.